APIGOV-21522 - Split the cache management for managedapp, accessreq and subscription to separate files
