+++ pkg/agent/cache/manager.go
@@ -559,150 +559,6 @@ func (c *cacheManager) GetSequence(watchTopicName string) int64 {
 	return 0
 }
 
-// ManagedApplication cache related methods
-func (c *cacheManager) GetManagedApplicationCacheKeys() []string {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	return c.managedApplicationMap.GetKeys()
-}
-
-func (c *cacheManager) AddManagedApplication(resource *v1.ResourceInstance) {
-	c.managedApplicationMap.SetWithSecondaryKey(resource.Metadata.ID, resource.Name, resource)
-}
-
-func (c *cacheManager) GetManagedApplication(id string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	managedApp, _ := c.managedApplicationMap.Get(id)
-	if managedApp != nil {
-		ri, ok := managedApp.(*v1.ResourceInstance)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-func (c *cacheManager) GetManagedApplicationByName(name string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	managedApp, _ := c.managedApplicationMap.GetBySecondaryKey(name)
-	if managedApp != nil {
-		ri, ok := managedApp.(*v1.ResourceInstance)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-func (c *cacheManager) DeleteManagedApplication(id string) error {
-	return c.managedApplicationMap.Delete(id)
-}
-
-// AccessRequest cache related methods
-func (c *cacheManager) GetAccessRequestCacheKeys() []string {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	return c.managedApplicationMap.GetKeys()
-}
-
-func (c *cacheManager) AddAccessRequest(ar *mv1.AccessRequest) {
-	appName := ar.Spec.ManagedApplication
-	instID := ""
-	for _, ref := range ar.Metadata.References {
-		if ref.Name == ar.Spec.ApiServiceInstance {
-			instID = ref.ID
-			break
-		}
-	}
-
-	instance, _ := c.GetAPIServiceInstanceByID(instID)
-	apiID := ""
-	apiStage := ""
-	if instance != nil {
-		apiID, _ = util.GetAgentDetailsValue(instance, defs.AttrExternalAPIID)
-		apiStage, _ = util.GetAgentDetailsValue(instance, defs.AttrExternalAPIStage)
-	}
-
-	c.accessRequestMap.SetWithSecondaryKey(ar.Metadata.ID, appName+":"+apiID+":"+apiStage, ar)
-}
-
-func (c *cacheManager) GetAccessRequestByAppAndAPI(appName, remoteAPIID, remoteAPIStage string) *mv1.AccessRequest {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	accessRequest, _ := c.accessRequestMap.GetBySecondaryKey(appName + ":" + remoteAPIID + ":" + remoteAPIStage)
-	if accessRequest != nil {
-		ri, ok := accessRequest.(*mv1.AccessRequest)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-func (c *cacheManager) GetAccessRequest(id string) *mv1.AccessRequest {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	accessRequest, _ := c.accessRequestMap.Get(id)
-	if accessRequest != nil {
-		ri, ok := accessRequest.(*mv1.AccessRequest)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-func (c *cacheManager) DeleteAccessRequest(id string) error {
-	return c.accessRequestMap.Delete(id)
-}
-
-// Subscription cache related methods - Temporary
-func (c *cacheManager) AddSubscription(resource *v1.ResourceInstance) {
-	if resource != nil {
-		c.subscriptionMap.SetWithSecondaryKey(resource.Metadata.ID, resource.Name, resource)
-	}
-}
-
-func (c *cacheManager) GetSubscriptionByName(subscriptionName string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	subscription, _ := c.subscriptionMap.GetBySecondaryKey(subscriptionName)
-	if subscription != nil {
-		ri, ok := subscription.(*v1.ResourceInstance)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-func (c *cacheManager) GetSubscription(id string) *v1.ResourceInstance {
-	c.ApplyResourceReadLock()
-	defer c.ReleaseResourceReadLock()
-
-	subscription, _ := c.subscriptionMap.Get(id)
-	if subscription != nil {
-		ri, ok := subscription.(*v1.ResourceInstance)
-		if ok {
-			return ri
-		}
-	}
-	return nil
-}
-
-func (c *cacheManager) DeleteSubscription(id string) error {
-	return c.subscriptionMap.Delete(id)
-}
-
 func (c *cacheManager) ApplyResourceReadLock() {
 	c.resourceCacheReadLock.Lock()
 }
